+++ index.js
@@ -1,9 +1,7 @@
 const fs = require('fs');
 const {promisify} = require('util');
-const ini = require('ini');
+const {parse} = require('./parser');
 const merge = require('lodash.merge');
-const {traverse} = require('traverse-async');
-const processNode = require('./lib/processNode');
 
 
 module.exports = (filename) => {
@@ -14,13 +12,18 @@ module.exports = (filename) => {
     return new Promise(async (resolve, reject) => {
         try {
             const contents = await readFile(filename, 'utf-8');
-            const parsed = ini.parse(contents);
+            const parsed = parse(contents);
             const environmentConfig = parsed[environment];
             const merged = merge(parsed.default, environmentConfig, {environment});
-            traverse(merged, processNode, (config) => {
-                return resolve(config);
-            });
+            return resolve(merged);
         } catch (error) {
+            if (error.location) {
+                const {start: {line, column}} = error.location;
+                const e = new Error(`INI parser failed '${error.message}'`);
+                e.original = error;
+                e.stack = `    at ${filename}:${line}:${column}\n${error.stack.split('\n').slice(1).join('\n')}`;
+                reject(e);
+            }
             reject(error);
         }
     });